+++ src/index.js
@@ -1,4 +1,5 @@
 import * as K from 'kefir'
+
 export * from 'kefir'
 
 /**
@@ -165,13 +166,39 @@ export function combinePropertyObject (props) {
 }
 
 
+/**
+ * Gets the current value of a property and throws an error if the
+ * property does not have a value.
+ *
+ * WARNING: Use this sparsely. Using this leads to un-idomatic code
+ *
+ * @param {K.Property<T>} prop
+ * @returns {T}
+ */
+export function getValue (prop) {
+  let called = false
+  let value
+  const off = onValue(prop, (x) => {
+    value = x
+    called = true
+  })
+
+  off()
+  if (!called) {
+    throw new Error('Property does not have current value')
+  }
+
+  return value
+}
+
+
 function assertIsProperty (prop) {
   if (
-    !prop ||
-    typeof prop.getType !== 'function'
+    !prop
+    || typeof prop.getType !== 'function'
     || prop.getType() !== 'property'
   ) {
-    throw new TypeError('Object values must be properties')
+    throw new TypeError(`Value ${prop} must be a property`)
   }
 }