+++ CHANGELOG.md
@@ -3,6 +3,7 @@ Changelog
 
 ## Upcoming
 
+* Added `K.getValue()` function.
 * `K.combinePropertyObject()` now returns a proper property
 
 ## v0.5.0 2016-10-16